YUK-21531 - Itron multiple enrollment not sent correctly
 - Updated enrollment code so that enrolling an Itron device in multiple groups causes Yukon to send an esiGroupRequest to Itron HCM for each group that the device is enrolled in.
 - Updated enrollment code so that updating enrollment of an Itron device causes Yukon to send an esiGroupRequest to Itron HCM for each group that the device was unenrolled from, even if it remains enrolled in some groups.

+++ yukon-client/common/src/com/cannontech/common/util/SftpConnection.java
@@ -97,6 +97,9 @@ public class SftpConnection implements AutoCloseable {
     
     @Override
     public void close() {
+        if (privateKeyFile == null) {
+            return;
+        }
         try {
             Files.delete(privateKeyFile.toPath());
         } catch (Exception e) {
